Renamed files as per revised naiming convention
