+++ public/index.html
@@ -8,7 +8,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>BookTrack - Add Book</title>
     <link rel="stylesheet" href="css/style.css">
-    <link rel="stylesheet" href="css/addBook.css">
+    <link rel="stylesheet" href="css/add-book.css">
 </head>
 <body>
     <div class="header">
@@ -87,7 +87,7 @@
     </div>
     
     <!-- Link to the external JavaScript file for handling 'Add Book' form interactions -->
-    <script src="js/addBook.js"></script>
+    <script src="js/add-book.js"></script>
 
     <!-- END FOR ADD-BOOK FUNCTION -->
